Forgot to commit

+++ src/main/java/view/service/IngredientService.java
@@ -27,12 +27,14 @@ public class IngredientService {
 
     public void addIngredient(){
         String name = inputHelper.getStringInput("Please enter the name of the ingredient");
-        ingredientRepository.addIngredient( name );
+        String unit = inputHelper.getStringInput("Please enter a unit of measurement");
+        ingredientRepository.addIngredient( name, unit );
     }
 
-    public Ingredient addIngredient(String name){
-        return ingredientRepository.addIngredient(name);
+    public Ingredient addIngredient(String name, String unit){
+        return ingredientRepository.addIngredient(name, unit);
     }
+
 
     public Ingredient findIngredientById(){
         int id = inputHelper.getIntegerInput("Enter the ID of the ingredient you would like to modify");
@@ -47,15 +49,19 @@ public class IngredientService {
         int id = inputHelper.getIntegerInput("Enter the ID of the ingredient you would like to delete");
         Ingredient ingredient = ingredientRepository.findIngredientById(id);
 
-        System.out.println("Delete " + ingredient.getIngredientName() + " ? Y/N ");
-        String input = inputHelper.getStringInput();
+        if(ingredient == null){
+            System.out.println("No ingredient found with the ID you entered.");
+        }else {
+            System.out.println("Delete " + ingredient.getIngredientName() + " ? Y/N ");
+            String input = inputHelper.getStringInput();
 
-        if( input.equalsIgnoreCase("y") ){
-            ingredientRepository.removeIngredient(ingredient);
-        }else if( input.equalsIgnoreCase("n")){
-            System.out.println( "Ingredient not deleted.");
-        }else{
-            System.out.println("Invalid input - please try again.");
+            if (input.equalsIgnoreCase("y")) {
+                ingredientRepository.removeIngredient(ingredient);
+            } else if (input.equalsIgnoreCase("n")) {
+                System.out.println("Ingredient not deleted.");
+            } else {
+                System.out.println("Invalid input - please try again.");
+            }
         }
     }
